move RootContext back to failgood package. it's not internal
