+++ failgood/jvm/test/failgood/experiments/given/ComplexGivenTest.kt
@@ -3,7 +3,7 @@
 package failgood.experiments.given
 
 import failgood.*
-import failgood.internal.RootContext
+import failgood.RootContext
 import strikt.api.expectThat
 import strikt.assertions.isEqualTo
