+++ failgood/jvm/test/failgood/functional/GivenTest.kt
@@ -3,7 +3,7 @@
 package failgood.functional
 
 import failgood.*
-import failgood.internal.RootContext
+import failgood.RootContext
 import strikt.api.expectThat
 import strikt.assertions.isEqualTo
