+++ failgood/jvm/test/failgood/internal/StringListTestFilterTest.kt
@@ -1,5 +1,6 @@
 package failgood.internal
 
+import failgood.RootContext
 import failgood.Test
 import failgood.describe
 import strikt.api.expectThat
